changed dataype to int

+++ lib/models/water_model.dart
@@ -1,6 +1,6 @@
 
 class WaterLog {
-  final double? amount;
+  final int? amount;
   final DateTime? time;
   WaterLog({this.amount, this.time});
 }
